% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nrel_limits}
\alias{nrel_limits}
\title{Limits used for NREL ocean uses analysis.}
\format{A nested list with the following items:
\describe{
  \itemize{
    \item{field}{: character of either "depth", "tide", "wave" or "wind"}
    \itemize{
      \item{breaks}{: numeric vector}
      \item{break_labels}{: weight of the diamond, in carats}
      \item{depth}{: "depth" present if field is "tide", "wave" or "wind"}
      \itemize{
        \item{min}{: numeric minimum depth}
        \item{max}{: numeric maximum depth}
      }
    }
  }
}}
\source{
\url{https://github.com/ecoquants/nrelutils}
}
\usage{
nrel_limits
}
\description{
A hierarchical list with limits used.
}
\keyword{datasets}
