% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_utils.R
\name{ply_map}
\alias{ply_map}
\title{Quick map of polygon layers}
\usage{
ply_map(ply, fld_color = "layer")
}
\arguments{
\item{ply}{}
}
\value{
init for efh assuming polygons with layers column, categorical
}
\description{
Quick map of polygon layers
}
