% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_utils.R
\name{ply_to_tifs}
\alias{ply_to_tifs}
\title{Polygons (ply) to raster(s) on filesystem (tif)}
\usage{
ply_to_tifs(x, y, ter, key, field = "one", by = NA, sfx = "epsg4326.tif")
}
\arguments{
\item{x}{sf polygon object}

\item{y}{raster object to use as template}

\item{ter}{territory name prefix}

\item{key}{output directory and used in prefix="{ter}_{key}"}

\item{by}{character. Name of column in sf object by which to aggregate layers. Default=NULL}

\item{sfx}{character. Suffix of output tif. Default="epsg4326.tif"}
}
\value{
return leaflet map
}
\description{
Polygons (ply) to raster(s) on filesystem (tif)
}
