% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_utils.R
\name{r_map}
\alias{r_map}
\title{Plot raster bisected around antimeridian [0,360] to [-180,180]}
\usage{
r_map(r, aggregate_factor = 10)
}
\arguments{
\item{r}{raster}

\item{aggregate_factor}{factor to aggregate}
}
\value{
leaflet map with bisected raster and graticule. Pan to get other half.
}
\description{
Plot raster bisected around antimeridian [0,360] to [-180,180]
}
