% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_utils.R
\name{raster_clip}
\alias{raster_clip}
\title{Clip raster with crop & mask}
\usage{
raster_clip(x, y)
}
\arguments{
\item{x}{raster object to be clipped}

\item{y}{raster object to clip by}
}
\value{
raster x with crop(x, y) and mask(x, y)
}
\description{
Clip raster with crop & mask
}
