% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_utils.R
\name{raster_unwrap}
\alias{raster_unwrap}
\title{Unwrap raster [0,360] to [-180,180]}
\usage{
raster_unwrap(r)
}
\arguments{
\item{raster}{raster object}
}
\value{
unwrapped raster
}
\description{
Unwrap raster [0,360] to [-180,180]
}
