% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_utils.R
\name{raster_wrap}
\alias{raster_wrap}
\title{Wrap raster from [-180,180] to [0,360]}
\usage{
raster_wrap(r)
}
\arguments{
\item{r}{raster object [-180,180]}
}
\value{
wrapped raster [0, 360]
}
\description{
Wrap raster from [-180,180] to [0,360]
}
