% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_utils.R
\name{sf_intersects}
\alias{sf_intersects}
\title{Return intersecting geometries}
\usage{
sf_intersects(a, b)
}
\arguments{
\item{a}{sf object to slice}

\item{b}{sf object to intersect}
}
\value{
sliced sf object `a` based on intersecting with geometrys from `b`,
vs `sf::st_intersects()` which returns a boolean matrix (`sparse = FALSE`)
or a sparse list (a boolean matrix (`sparse = FALSE`) containing indices matched
in `b`.
}
\description{
Return intersecting geometries
}
