% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_utils.R
\name{sf_unwrap}
\alias{sf_unwrap}
\title{Unwrap sf geometry [0,360] to [-180,180]}
\usage{
sf_unwrap(x)
}
\arguments{
\item{x}{sf object [0,360]}
}
\value{
unwrapped sf object [-180,180] but could be outside range? (TODO: check!). The geometry column gets renamed to geom.
}
\description{
Unwrap sf geometry [0,360] to [-180,180]
}
