% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_utils.R
\name{sf_wrap}
\alias{sf_wrap}
\title{Wrap sf geometry [-180,180] to [0,360]}
\usage{
sf_wrap(x)
}
\arguments{
\item{x}{sf object [-180,180]}
}
\value{
wrapped sf object [0,360]. The geometry column gets renamed to geom.
}
\description{
Wrap sf geometry [-180,180] to [0,360]
}
