% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_utils.R
\name{sf_wrap_intersection}
\alias{sf_wrap_intersection}
\title{Wrap and intersect}
\usage{
sf_wrap_intersection(x, y)
}
\arguments{
\item{x}{sf object to intersect, eg layer}

\item{y}{sf object to intersect, eg eez}
}
\value{
wrapped and intersected sf object with a one field as a valid MULTIPOLYGON
}
\description{
Wrap and intersect
}
